Reviewing PositiveFilter.
