+++ src/exercise2/Mapping.java
