+++ src/exercise4/NumberMap.java
